requirements.txt
asgiref==3.5.1
Django==3.2.13
djangorestframework==3.13.1
importlib-metadata==4.11.3
install==1.3.5
Markdown==3.3.6
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
zipp==3.8.0
